Feat: patchInterviewVisibility serverAction으로 변경
