+++ src/libs/actions/interview.ts
@@ -1,5 +1,7 @@
 'use server';
 
+import { cookies } from 'next/headers';
+
 import { IRequestInterviewForm } from '@/types/interview';
 import {
   IResponseGetUploadUrl,
@@ -7,6 +9,8 @@ import {
 } from '@/types/Response/interview';
 import { apiServer } from '@/utils/apiServer';
 
+import { IPatchInterviewVisibilityPayload } from '../types/payload';
+import { IPatchInterviewVisibility } from '../types/response';
 import { reissueAccessToken } from './auth';
 
 export const getUploadUrl = async (
@@ -60,4 +64,33 @@ export const patchInterviewStatus = async (
   if (!ok) {
     throw new Error(`HTTP error! status: ${status}`);
   }
+};
+
+export const patchInterviewVisibility = async (
+  interviewId: number,
+  body: IPatchInterviewVisibilityPayload[],
+): Promise<IPatchInterviewVisibility> => {
+  const response = await apiServer.patch(
+    `api/v1/interviews/${interviewId}/visibility`,
+    {
+      body: JSON.stringify(body),
+      cache: 'no-store',
+    },
+  );
+
+  if (response.status === 401) {
+    return reissueAccessToken(() =>
+      patchInterviewVisibility(interviewId, body),
+    );
+  }
+
+  if (!response.ok) {
+    const errorData = await response.json();
+    cookies().set('errorMessage', errorData.errorMessage, {
+      maxAge: 5,
+    });
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+
+  return response.json();
 };
