+++ src/libs/services/interview.ts
@@ -1,15 +1,10 @@
 import { notFound } from 'next/navigation';
 
 import { IResponseGetInterview } from '@/types/Response/interview';
-import { apiClient } from '@/utils/apiClient';
 import { apiServer } from '@/utils/apiServer';
 
 import { reissueAccessToken } from '../actions/auth';
-import { IPatchInterviewVisibilityPayload } from '../types/payload';
-import {
-  IGetInterviewInformation,
-  IPatchInterviewVisibility,
-} from '../types/response';
+import { IGetInterviewInformation } from '../types/response';
 
 export const getInterviewInfo = async (
   interviewId: string,
@@ -58,25 +53,6 @@ export const putUploadMediaBlob = async (
   if (!response.ok) {
     throw new Error(`HTTP error! status: ${response.status}`);
   }
-};
-
-export const patchInterviewVisibility = async (
-  interviewId: number,
-  body: IPatchInterviewVisibilityPayload[],
-): Promise<IPatchInterviewVisibility> => {
-  const response = await apiClient.patch(
-    `api/v1/interviews/${interviewId}/visibility`,
-    {
-      body: JSON.stringify(body),
-      cache: 'no-store',
-    },
-  );
-
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-
-  return response.json();
 };
 
 export const getInterviewResult = async (
